chore: refactor decentralized handler
