+++ internal/node/component/decentralized/handler_workers.go
@@ -9,32 +9,36 @@ import (
 	"github.com/labstack/echo/v4"
 	"github.com/rss3-network/node/config"
 	"github.com/rss3-network/node/schema/worker"
+	"github.com/rss3-network/protocol-go/schema/network"
+	"github.com/rss3-network/protocol-go/schema/tag"
 )
+
+type WorkerResponse struct {
+	Data []*WorkerInfo `json:"data"`
+}
+
+type WorkerInfo struct {
+	Network  network.Network `json:"network"`
+	Worker   worker.Worker   `json:"worker"`
+	Tags     []tag.Tag       `json:"tags"`
+	Platform worker.Platform `json:"platform"`
+	Status   worker.Status   `json:"status"`
+}
 
 func (c *Component) GetWorkers(ctx echo.Context) error {
 	var wg sync.WaitGroup
 
-	workerInfoChan := make(chan WorkerInfo, len(c.config.Component.Decentralized))
+	workerCount := len(c.config.Component.Decentralized)
 
-	workers := make([]WorkerInfo, 0, len(c.config.Component.Decentralized))
+	workerInfoChan := make(chan *WorkerInfo, workerCount)
 
 	for _, w := range c.config.Component.Decentralized {
 		wg.Add(1)
 
-		go func(w *config.Module) {
+		go func(module *config.Module) {
 			defer wg.Done()
 
-			status := c.getWorkerStatus(ctx.Request().Context(), w.Network.String(), w.Worker.String())
-
-			workerInfo := WorkerInfo{
-				Network:  w.Network,
-				Worker:   w.Worker,
-				Platform: worker.ToPlatformMap[w.Worker],
-				Tags:     worker.ToTagsMap[w.Worker],
-				Status:   status,
-			}
-
-			workerInfoChan <- workerInfo
+			workerInfoChan <- c.fetchWorkerInfo(ctx.Request().Context(), module)
 		}(w)
 	}
 
@@ -44,6 +48,7 @@ func (c *Component) GetWorkers(ctx echo.Context) error {
 		close(workerInfoChan)
 	}()
 
+	workers := make([]*WorkerInfo, 0, workerCount)
 	for workerInfo := range workerInfoChan {
 		workers = append(workers, workerInfo)
 	}
@@ -53,7 +58,7 @@ func (c *Component) GetWorkers(ctx echo.Context) error {
 	})
 }
 
-// getWorkerStatus gets worker status from Redis cache by network and workName.
+// getWorkerStatus gets worker status from Redis cache by network and workerName.
 func (c *Component) getWorkerStatus(ctx context.Context, network, workerName string) worker.Status {
 	if c.redisClient == nil {
 		return worker.StatusUnknown
@@ -78,6 +83,19 @@ func (c *Component) getWorkerStatus(ctx context.Context, network, workerName str
 	}
 
 	return status
+}
+
+func (c *Component) fetchWorkerInfo(ctx context.Context, module *config.Module) *WorkerInfo {
+	// Fetch status from a specific worker network.
+	status := c.getWorkerStatus(ctx, module.Network.String(), module.Worker.String())
+
+	return &WorkerInfo{
+		Network:  module.Network,
+		Worker:   module.Worker,
+		Platform: worker.ToPlatformMap[module.Worker],
+		Tags:     worker.ToTagsMap[module.Worker],
+		Status:   status,
+	}
 }
 
 // buildCacheKey builds cache key for Redis.
